Fixed the wrong data types in the filterbank header.

+++ include/config.hpp
@@ -53,7 +53,7 @@ inline void default_config(config_s &config) {
     config.chunks = 32;
     config.fftsize = 32;
     config.freqavg = 9;
-    config.foff = 1/27 * config.freqavg;
+    config.foff = (double)1.0/(double)27.0 * (double)config.freqavg;
     config.gulp = 16384;        // 2^24, equivalent to ~1.75s for 108us sampling time (for testing purposes)
     //config.gulp = 131072;       // 2^17, equivalent to ~14s for 108us sampling time
     config.nchans = 7;
